requirements.txt
pyrender==0.1.45
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.5
scikit-image==0.18.1
trimesh==3.9.9
scipy==1.5.1
Pillow==7.2.0
imageio==2.9.0
numpy==1.19.5
structlog==21.1.0
matplotlib==3.3.4
tqdm==4.59.0
imgaug==0.4.0
opencv-python==4.5.1.48
# torch==1.8.0+cu110
# torchvision==0.9.0+cu110
#pip install torch==1.8.0+cu111 torchvision==0.9.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html
PyYAML==5.4.1
tensorboard==2.4.1
#Blender==2.80
#cudatoolkit==11.0
# detectron2==0.4+cu110
#python -m pip install detectron2 -f https://dl.fbaipublicfiles.com/detectron2/wheels/cu111/torch1.8/index.html
# pytorch3d==0.4.0
#pip install "git+https://github.com/facebookresearch/pytorch3d.git"


### BOP TOOLKIT
scipy==1.5.1
kiwisolver==1.3.1
#matplotlib==2.2.4
#imageio==2.5.0
pypng==0.0.19
#Cython==0.29.24
#PyOpenGL==3.1.0
triangle==20190115.2
glumpy==1.1.0
#opencv-python==4.3.0.36
#Pillow>=8.2.0
git+https://github.com/MartinSmeyer/cocoapi.git@v1.0#subdirectory=PythonAPI
vispy>=0.6.5

# Self added for BOP
pytorch3d
